#12151: Rename max_pool2d to generic pool op
